+++ actix-web-codegen/CHANGES.md
@@ -1,7 +1,10 @@
 # Changes
 
 ## Unreleased - 2021-xx-xx
-- Preserve function visibility for routing macros
+- Preserve function visibility for routing macros. [#2714]
+
+[#2714]: https://github.com/actix/actix-web/pull/2714
+
 
 ## 4.0.0 - 2022-02-24
 - Version aligned with `actix-web` and will remain in sync going forward.