+++ actix-web-codegen/CHANGES.md
@@ -1,8 +1,10 @@
 # Changes
 
 ## Unreleased - 2021-xx-xx
-### Fixed
-- Allow a path as a guard in route handler macro, such as `get`, `post`, as the `actix-web` document says
+- Fix support for guard paths in route handler macros. [#2771]
+
+[#2771] https://github.com/actix/actix-web/pull/2771
+
 
 ## 4.0.0 - 2022-02-24
 - Version aligned with `actix-web` and will remain in sync going forward.